feat: use ffi to check vaapi capabilities (#1)
* refactor: remove unneeded partialeq

* feat: use ffi to check vaapi capabilities

* feat: more vaapi profiles for capabilities checks

* feat: more vaapi profiles for capabilities checks

* fix: vaapi software decode and hardware encode

+++ crates/ffpipeline/src/pipeline.rs
@@ -83,7 +83,7 @@ impl PixelFormat {
         }
     }
 
-    pub(crate) fn bit_depth(&self) -> u32 {
+    pub(crate) fn bit_depth(&self) -> u8 {
         match self {
             PixelFormat::Yuv420p | PixelFormat::Nv12 => 8,
             PixelFormat::Yuv420p10le | PixelFormat::P010le => 10,
